Merge branch 'S_master1'

+++ htdocs/view/blocks.php
@@ -75,6 +75,10 @@ $institution = $view->get('institution');
 
 // check if cancel was selected
 if ($new && isset($_POST['cancel'])) {
+    if ($view->get('type') == 'profile' || $view->get('type') == 'dashboard') {
+        throw new AccessDeniedException(get_string('cantdeleteview', 'view'));
+    }
+    form_validate(param_variable('sesskey', null));
     $view->delete();
     $view->post_edit_redirect();
 }
@@ -148,7 +152,7 @@ foreach (array_reverse($THEME->get_url('style/style.css', true, 'artefact/file')
 // longer available to them.
 if ($viewtheme && !isset($allowedthemes[$viewtheme])) {
     $smarty = smarty(array(), $stylesheets, false, $extraconfig);
-    $smarty->assign('maintitle', hsc(TITLE));
+    $smarty->assign('maintitle', TITLE);
     $smarty->assign('formurl', get_config('wwwroot') . 'view/blocks.php');
     $smarty->assign('view', $view->get('id'));
     $smarty->assign('viewtitle', $view->get('title'));
@@ -187,7 +191,7 @@ foreach (array_keys($_POST + $_GET) as $key) {
     }
 }
 
-$smarty->assign('maintitle', hsc(TITLE));
+$smarty->assign('maintitle', TITLE);
 $smarty->assign('displaylink', $displaylink);
 $smarty->assign('formurl', get_config('wwwroot') . 'view/blocks.php');
 $smarty->assign('category', $category);
@@ -229,7 +233,6 @@ if (get_config('viewmicroheaders')) {
     }
 }
 
-$smarty->assign('userdisplayname', display_name($USER, null, true));
 $smarty->assign('viewtype', $viewtype);
 $smarty->assign('view', $view->get('id'));
 $smarty->assign('groupid', $group);
